tee/test: created generic enclave test.
- Refactored prototype enclave test to tee/test.
- Promoted tee/prototype.ErrEnclaveStopped to tee.ErrEnclaveStopped.
- Moved tee/test.NewRandomAddress to eth.NewRandomAddress to break
  cyclic dependencies.
